version123

+++ app.py
@@ -40,73 +40,73 @@ def flask_app(host=None, port=None):
 
 
   def data_preparation():
-    
-    if os.getenv("FLASK_ENV") == "development":
-      logging.info("Entered the if block###########################")
-      load_dotenv()
-      database_url=os.getenv("DATABASE_URL")
-      
-      with open("private_key.txt", "r") as file:
-        private_key = file.read()
-
-      # Explicitly decode the private key using UTF-8
-      private_key = private_key.encode('utf-8').decode('unicode_escape')
-      
-      client_email = os.getenv("CLIENT_EMAIL")
-    else:
-      logging.info("Entered the else block###########################")
-      # Retrieve the private key from the environment variable
-      private_key_str = os.environ.get('PRIVATE_KEY')
-      logging.info(f"PRIVATE_KEY: {os.environ.get('PRIVATE_KEY')}")
-      # Replace the escaped newline sequences with actual newlines
-      private_key = private_key_str.encode('utf-8').decode('unicode_escape')
-      logging.info(f"Decoded Private Key: {private_key}")
-
-      client_email = os.environ.get('CLIENT_EMAIL')
-      database_url = os.environ.get('DATABASE_URL')
-    
-    with psycopg2.connect(database_url) as connection:
-      sql_query2 = 'SELECT * FROM "questions_potentialcustomers"'
-      df_potential_customer = pd.read_sql(sql_query2, connection)
-      sql_query = 'SELECT * FROM "order_existing_clients"'
-      df_existing_customer_original = pd.read_sql(sql_query, connection)
-      
-      with connection.cursor() as cursor:
-          # Execute the SQL query
-          sql_query = 'SELECT * FROM "orders"'
-          cursor.execute(sql_query)
-
-          # Fetch all rows
-          rows = cursor.fetchall()
-
-    # Extract column names
-    columns = [desc[0] for desc in cursor.description]
-
-    # Format data as a string
-    data_rows = []
-    for row in rows:
-        data_rows.append(" | ".join(map(str, row)))
-
-    # Create the final string
-    df_existing_customer = " | ".join(columns) + "\n" + "\n".join(data_rows)
-
-
-    
-    #DOWNLOAD AND CREATE THE WORD FILE
-    project_id = "deployment-391914"
-    private_key_id = "2de8528d9202cd246961502047094035cfd851fb"
-    client_id = "104118424303777600500"
-    auth_uri = "https://accounts.google.com/o/oauth2/auth"
-    token_uri = "https://oauth2.googleapis.com/token"
-    auth_provider_x509_cert_url = "https://www.googleapis.com/oauth2/v1/certs"
-    client_x509_cert_url = "https://www.googleapis.com/robot/v1/metadata/x509/textfile-aichatbot%40deployment-391914.iam.gserviceaccount.com"
-    universe_domain = "googleapis.com"
-
-
-    SCOPES = ['https://www.googleapis.com/auth/drive.readonly']
-    
-    credentials = None
     try:
+      if os.getenv("FLASK_ENV") == "development":
+        
+        load_dotenv()
+        database_url=os.getenv("DATABASE_URL")
+        
+        with open("private_key.txt", "r") as file:
+          private_key = file.read()
+
+        # Explicitly decode the private key using UTF-8
+        private_key = private_key.encode('utf-8').decode('unicode_escape')
+        
+        client_email = os.getenv("CLIENT_EMAIL")
+      else:
+      
+        # Retrieve the private key from the environment variable
+        private_key_str = os.environ.get('PRIVATE_KEY')
+        logging.info(f"PRIVATE_KEY: {os.environ.get('PRIVATE_KEY')}")
+        # Replace the escaped newline sequences with actual newlines
+        private_key = private_key_str.encode('utf-8').decode('unicode_escape')
+        logging.info(f"Decoded Private Key: {private_key}")
+
+        client_email = os.environ.get('CLIENT_EMAIL')
+        database_url = os.environ.get('DATABASE_URL')
+    
+      with psycopg2.connect(database_url) as connection:
+        sql_query2 = 'SELECT * FROM "questions_potentialcustomers"'
+        df_potential_customer = pd.read_sql(sql_query2, connection)
+        sql_query = 'SELECT * FROM "order_existing_clients"'
+        df_existing_customer_original = pd.read_sql(sql_query, connection)
+        
+        with connection.cursor() as cursor:
+            # Execute the SQL query
+            sql_query = 'SELECT * FROM "orders"'
+            cursor.execute(sql_query)
+
+            # Fetch all rows
+            rows = cursor.fetchall()
+
+      # Extract column names
+      columns = [desc[0] for desc in cursor.description]
+
+      # Format data as a string
+      data_rows = []
+      for row in rows:
+          data_rows.append(" | ".join(map(str, row)))
+
+      # Create the final string
+      df_existing_customer = " | ".join(columns) + "\n" + "\n".join(data_rows)
+
+
+    
+      #DOWNLOAD AND CREATE THE WORD FILE
+      project_id = "deployment-391914"
+      private_key_id = "2de8528d9202cd246961502047094035cfd851fb"
+      client_id = "104118424303777600500"
+      auth_uri = "https://accounts.google.com/o/oauth2/auth"
+      token_uri = "https://oauth2.googleapis.com/token"
+      auth_provider_x509_cert_url = "https://www.googleapis.com/oauth2/v1/certs"
+      client_x509_cert_url = "https://www.googleapis.com/robot/v1/metadata/x509/textfile-aichatbot%40deployment-391914.iam.gserviceaccount.com"
+      universe_domain = "googleapis.com"
+
+
+      SCOPES = ['https://www.googleapis.com/auth/drive.readonly']
+      
+      credentials = None
+      
       credentials = service_account.Credentials.from_service_account_info(
         {
           "type": "service_account",
@@ -126,35 +126,46 @@ def flask_app(host=None, port=None):
 
       # Use credentials to make API requests
 
-    except exceptions.GoogleAuthError as e:
-      print(f"Error creating credentials: {e}")
+   
       
+      if credentials:
+      # Use credentials to make API requests
+        file_id = '152GW4g2WrNjGeaFuhP7-RCX7YWDPM4GE'
+        service = build('drive', 'v3', credentials=credentials)
+        request = service.files().get_media(fileId=file_id)
 
-    file_id = '152GW4g2WrNjGeaFuhP7-RCX7YWDPM4GE'
-    service = build('drive', 'v3', credentials=credentials)
-    request = service.files().get_media(fileId=file_id)
+        temp_folder = os.path.join(tempfile.gettempdir(), 'my_temp_folder')
+        os.makedirs(temp_folder, exist_ok=True)
+        file_path = os.path.join(temp_folder, 'Cars_services_downloaded.docx')
 
-    temp_folder = os.path.join(tempfile.gettempdir(), 'my_temp_folder')
-    os.makedirs(temp_folder, exist_ok=True)
-    file_path = os.path.join(temp_folder, 'Cars_services_downloaded.docx')
+        with open(file_path, 'wb') as f:
+          downloader = MediaIoBaseDownload(f, request)
+          done = False
+          while done is False:
+              status, done = downloader.next_chunk()
+    
 
-    with open(file_path, 'wb') as f:
-        downloader = MediaIoBaseDownload(f, request)
-        done = False
-        while done is False:
-            status, done = downloader.next_chunk()
+    
 
-    # Read the content of the Word document using python-docx
-    doc = Document(file_path)
+        # Read the content of the Word document using python-docx
+        doc = Document(file_path)
 
-    full_text=""
-    for paragraph in doc.paragraphs:
-      full_text+=paragraph.text+ "\n"
-    word_text=full_text.strip()
+        full_text=""
+        for paragraph in doc.paragraphs:
+          full_text+=paragraph.text+ "\n"
+        word_text=full_text.strip()
 
-    os.remove(file_path)
+        os.remove(file_path)
 
-    return df_existing_customer_original, df_existing_customer, df_potential_customer, word_text
+        return df_existing_customer_original, df_existing_customer, df_potential_customer, word_text
+    except exceptions.GoogleAuthError as e:
+          # Log the exception for debugging purposes
+          logging.error(f"Error creating credentials: {e}")
+          return None
+    except Exception as e:
+        # Log the exception for debugging purposes
+        logging.error(f"An error occurred: {e}")
+        return None
 
 #-----------------------------------------------------------------------------------------------
 #             Important variables
